feat: model interface change.
