+++ src/strands/experimental/bidirectional_streaming/__init__.py
@@ -4,7 +4,7 @@
 from .agent.agent import BidirectionalAgent
 
 # Advanced interfaces (for custom implementations)
-from .models.bidirectional_model import BidirectionalModel, BidirectionalModelSession
+from .models.base_model import BidirectionalModel, BidirectionalModelSession
 
 # Model providers - What users need to create models
 from .models.novasonic import NovaSonicBidirectionalModel
