+++ src/strands/experimental/bidirectional_streaming/models/__init__.py
@@ -1,6 +1,6 @@
 """Bidirectional model interfaces and implementations."""
 
-from .bidirectional_model import BidirectionalModel, BidirectionalModelSession
+from .base_model import BidirectionalModel, BidirectionalModelSession
 from .gemini_live import GeminiLiveBidirectionalModel, GeminiLiveSession
 from .novasonic import NovaSonicBidirectionalModel, NovaSonicSession
 from .openai import OpenAIRealtimeBidirectionalModel, OpenAIRealtimeSession
