+++ src/strands/experimental/bidirectional_streaming/models/novasonic.py
@@ -37,7 +37,7 @@ from ..types.bidirectional_streaming import (
     TextOutputEvent,
     UsageMetricsEvent,
 )
-from .bidirectional_model import BidirectionalModel, BidirectionalModelSession
+from .base_model import BaseModel
 
 logger = logging.getLogger(__name__)
