+++ src/strands/experimental/bidirectional_streaming/models/openai.py
@@ -26,7 +26,7 @@ from ..types.bidirectional_streaming import (
     TextOutputEvent,
     VoiceActivityEvent,
 )
-from .bidirectional_model import BidirectionalModel, BidirectionalModelSession
+from .base_model import BaseModel
 
 logger = logging.getLogger(__name__)
